reorganize files into proper server project layout
